move conftest to tests folder, add demo scene restoration fixture, fix tests
