+++ tests/functional/test_camera_deletion_api.py
@@ -135,7 +135,7 @@ class CameraDeletionTest(FunctionalTest):
 
     return
 
-def test_camera_deletion_main(scenescape_env, request, record_xml_attribute):
+def test_camera_deletion_main(scenescape_env, demo_scene, request, record_xml_attribute):
   test = CameraDeletionTest(TEST_NAME, request, record_xml_attribute)
   test.testCameraDeletion()
   assert test.exitCode == 0
