+++ tests/functional/test_rest_test.py
@@ -242,7 +242,7 @@ class TestAPI(FunctionalTest):
       self.recordTestResult()
     return
 
-def test_api(scenescape_env, request, record_xml_attribute):
+def test_api(scenescape_env, demo_scene, request, record_xml_attribute):
   test = TestAPI(TEST_NAME, request, record_xml_attribute)
   test.verifyThings()
   assert test.exitCode == 0
