+++ tests/pytest.ini
@@ -3,11 +3,11 @@
 
 [pytest]
 testpaths =
-    tests/functional
-    tests/ui
-    tests/security/system
-    tests/system/stability
-    tests/sscape_tests
+    functional
+    ui
+    security/system
+    system/stability
+    sscape_tests
 
 addopts = -v --tb=short -s
